requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.12.0
openai-agents>=0.3.1  # OpenAI Agents SDK for MAI-DxO implementation
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Evaluation
langfuse>=2.0.0

# Logging and monitoring
python-json-logger>=2.0.7
rich>=13.7.0

# Utilities
httpx>=0.25.0
tenacity>=8.2.3
pyyaml>=6.0.1
python-multipart>=0.0.6

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# MCP SDKs (if available as packages)
# ref-tools
# exa-py